2D Model optimisation (#415)
* 2D Model optimisation

* whats new

+++ oggm/tests/test_prepro.py
@@ -426,8 +426,8 @@ class TestCenterlines(unittest.TestCase):
         nb = len(np.where(rest == 2)[0])
         nc = len(np.where(rest == 1)[0])
         nd = len(np.where(rest == 0)[0])
-        denom = np.float64((na+nc)*(nd+nc)+(na+nb)*(nd+nb))
-        hss = np.float64(2.) * ((na*nd)-(nb*nc)) / denom
+        denom = np.float((na+nc)*(nd+nc)+(na+nb)*(nd+nb))
+        hss = np.float(2.) * ((na*nd)-(nb*nc)) / denom
         if cfg.PARAMS['grid_dx_method'] == 'linear':
             self.assertTrue(hss > 0.53)
         if cfg.PARAMS['grid_dx_method'] == 'fixed':  # quick fix
